third_party/blink/web_tests/external/wpt/loading/lazyload/image-loading-lazy-load-event.tentative-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: In-viewport loading=lazy images do not block the window load event
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
FAIL In-viewport loading=lazy images do not block the window load event assert_true: The window load event should fire before the in-viewport loading=lazy images load expected true got false
Harness: the test ran to completion.
